валидация формы, исправления, разбивка на модули
