+++ js/form-validation.js
@@ -1,45 +1,6 @@
 import {checkStringLength} from './util.js';
 
-//работа с формой
-const bodyDocument = document.querySelector('body');
-const uploadFile = document.querySelector('#upload-file');
-const uploadForm = document.querySelector('#upload-select-image');
-const uploadFileOverlay = document.querySelector('.img-upload__overlay');
-const buttonClose = document.querySelector('#upload-cancel');
-const hashtagField = document.querySelector('.text__hashtags');
-const descriptionField = document.querySelector('.text__description');
-
-//функция переключения классов для открытия или закрытия модального окна
-const OpenOrCloseModal =  () => {
-  uploadFileOverlay.classList.toggle('hidden');
-  bodyDocument.classList.toggle('modal-open');
-};
-
-//отмена закрытия формы по эскейпу при фокусе в поле ввода
-hashtagField.addEventListener('keydown', (evt) => {
-  evt.stopPropagation();
-});
-descriptionField.addEventListener('keydown', (evt) => {
-  evt.stopPropagation();
-});
-
-//открытие формы редактирования изображения после загрузки файла
-uploadFile.addEventListener('change', () => {
-  OpenOrCloseModal();
-});
-
-//закрытие формы редактирования изображения
-buttonClose.addEventListener('click', () => {
-  OpenOrCloseModal();
-  uploadForm.reset();
-});
-
-document.addEventListener('keydown', (evt) => {
-  if (evt.key === 'Escape') {
-    OpenOrCloseModal();
-    uploadForm.reset();
-  }
-});
+//модуль отвечает за валидацию формы загрузки нового изображения (поля хэштегов и описания)
 
 //константы
 const HASHTAG_MAX_AMOUNT = 5;
@@ -52,8 +13,11 @@ const errorMessage = {
   DESCRIPTION_LENGTH : `Максимальная длина описания - ${DESCRIPTION_MAX_LENGTH} символов`
 };
 
-//Валидация хэштегов и описания
+//Валидируемая форма
 const imgUploadForm = document.querySelector('.img-upload__form');
+const hashtagField = document.querySelector('.text__hashtags');
+const descriptionField = document.querySelector('.text__description');
+
 
 const pristine = new Pristine(imgUploadForm, {
   classTo: 'img-upload__field-wrapper',
