+++ js/main.js
@@ -1,5 +1,6 @@
 import {genPosts} from './data.js';
 import {getPictures} from './pictures.js';
+import './modal.js';
 import './form-validation.js';
 
 
@@ -8,3 +9,4 @@ const posts = genPosts();
 
 //прорисовка постов на сайте
 getPictures(posts);
+
